requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

openai
python-dotenv
azure-core
azure-ai-documentintelligence
azure-functions
azure-storage-blob
azure-identity
azure-keyvault-secrets
langchain
tiktoken
azure-search-documents

# Apartir desde aqui a abajo son dependencias que son utilizadas para pruebas, las cuales no deben ser instaladas en produccion.
flask